requirements.txt
appdirs==1.4.4
attrs==21.2.0
black==21.7b0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
flake8==3.9.2
google-api-core==1.31.2
google-auth==1.35.0
google-cloud-bigtable==1.7.0
google-cloud-core==1.7.2
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
idna==3.2
iniconfig==1.1.1
isort==5.9.3
libcst==0.3.20
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pluggy==0.13.1
proto-plus==1.19.0
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
